moves state related methods into seperate file, keeps it within fipslookup module
